🚧🧑‍💻 (Typescript) - Rewrite sidebar.js to sidebar.ts / All functions are in the same files, so they need to be improved
